Issue #12 Reverse Translation
implementation for reverse translation as pet issue #12. What I do here
is that i try to generate all possible permutations that could have
rendered the given piglatin word and give it as a List<String> rather
thant just one String. Similarly, if given a sentence, we will make "a
tree" of all possible combinations as a List<Strings>. Does this make
sense to you guys?